backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
websockets==12.0

# Database
sqlmodel==0.0.14
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# ML/AI
onnxruntime==1.16.3
opencv-python-headless==4.9.0.80
pillow==10.2.0
numpy==1.26.3

# Reports & Export
reportlab==4.0.9
pandas==2.2.0

# Utils
python-dateutil==2.8.2
pytz==2024.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Code Quality
black==24.1.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2

# Redis (optional)
redis==5.0.1
hiredis==2.3.2

# Web Push
py-vapid==1.9.0
pywebpush==1.14.0
